+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * issue: Ensure that panels with links use callbacks
 * issue: Vertical scroll bar was not appearing on some themes
 * feature#208: Add select all to permissions tab
+* feature: Add DNS resolving check
 
 --- 4.0.1 ---